chore: Moves spec files to the src folder - iteration 5
