--- superset-frontend/spec/javascripts/components/CachedLabel_spec.tsx
@@ -1,42 +1,0 @@
-/**
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-import React from 'react';
-import { render, screen } from 'spec/helpers/testing-library';
-
-import CachedLabel, { CacheLabelProps } from 'src/components/CachedLabel';
-
-const defaultProps = {
-  onClick: () => {},
-  cachedTimestamp: '2017-01-01',
-};
-
-const setup = (props: CacheLabelProps) => <CachedLabel {...props} />;
-
-describe('CachedLabel', () => {
-  it('is valid', () => {
-    expect(React.isValidElement(<CachedLabel {...defaultProps} />)).toBe(true);
-  });
-
-  it('renders', () => {
-    render(setup(defaultProps));
-
-    const label = screen.getByText(/cached/i);
-    expect(label).toBeVisible();
-  });
-});
